More efficient two body integrals
The trapezoidal method applied to a vector can be implemented as multiplying it by dx, halving the ends, and taking the sum of the elements.

Since we are applying the trapezoidal method to a product of 3 vectors, we can perform the first two steps on the vector which is available outside the double loop. We can then perform the first vector product inside the first loop, and only perform a dot product inside the double loop.

This is much faster than computing two vector products and applying the trapezoidal method inside the double loop.